all: 进行代码可读性修改

+++ log/log.go
@@ -2,10 +2,10 @@
 package log
 
 const (
-	RecordFullType   = uint8(1)
-	RecordFirstType  = uint8(2)
-	RecordMiddleType = uint8(3)
-	RecordLastType   = uint8(4)
+	RecordFullType   = 1
+	RecordFirstType  = 2
+	RecordMiddleType = 3
+	RecordLastType   = 4
 
 	BlockSize  = 32768
 	HeaderSize = 7 // 4 (checksum) + 2 (length) + 1 (type)
